requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
passlib[bcrypt]
python-jose
redis
pydantic
asyncpg